+++ lib/portfolio/presenter/controller/wallet_controller.dart
@@ -1,7 +1,7 @@
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
-import '../../data/model/coin_model.dart';
-import '../../data/repository/coin_repository.dart';
+import '../../model/coin_model.dart';
+import '../../repository/coin_repository.dart';
 
 class WalletController extends ChangeNotifier {
   CoinRepository coinRepo;
