+++ lib/portfolio/presenter/ui/home_page/widgets/crypto_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../../shared/provider/provider.dart';
+import '../../../../../shared/provider/provider.dart';
 import 'crypto_item.dart';
 
 class CryptoList extends HookConsumerWidget {
