+++ lib/portfolio/presenter/ui/home_page/widgets/wallet_header.dart
@@ -5,7 +5,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
 import '../../../controller/wallet_controller.dart';
 
-import '../../../../shared/provider/provider.dart';
+import '../../../../../shared/provider/provider.dart';
 
 class WalletHeader extends HookConsumerWidget {
   const WalletHeader({
