+++ lib/shared/provider/provider.dart
@@ -1,7 +1,7 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../data/repository/coin_repository.dart';
-import '../../presenter/controller/wallet_controller.dart';
+import '../../portfolio/repository/coin_repository.dart';
+import '../../portfolio/presenter/controller/wallet_controller.dart';
 
 final viewWalletValueProvider = StateProvider<bool>(
   (ref) => false,
